Exynos: "tickless idle" configurable with default=n
With "tickless idle", a CPU core executing the idle_op() function
disables its timer interrupt and either goes into the halt state or
even disables itself (it will be woken up by an external interrupt
in that case).

Currently there is a problem when the boot CPU disables its timer
interrupts: The KIP clock is no longer updated (it's only updated
by the boot CPU) and a timeout on another CPU is never hit (because
the function checking that also uses the KIP clock).

Hence, we disable that mode by default.

See ticket #CD-387.

Change-Id: I854b8d108c41fead25f7a00b5338b7a2025ea60b