add fiasco.image to KERNEL_EXTRA
On x86, 'fiasco' and 'fiasco.debug' are targets of separate subsystems
(BOOT versus KERNEL). Hence, on 'make clean', both files are removed.
On all other systems, 'fiasco.debug' is created as part of the KERNEL
subsystem. Add 'fiasco.debug' to KERNEL_EXTRA on these platforms so it
gets removed on 'make clean' as well.

This enforces a consistent behavior between x86 and the other platforms
and IMO it does not make sense to keep 'fiasco.debug' when 'fiasco' is
removed on 'make clean'.

Change-Id: I03e4244d6d322137e5640677c878a7312fc47b44