jdb: rework capability dump call
Many entities write testing/debugging information during testruns and
want to have them passed to analysis tools. To unify this, a common
format for guard lines for such blocks of information was decided upon.
Since the capability dump feature is one such entity, its guard lines
need to be adapted, which this commit does.

Additionally, it makes the dump call more robust to the absence of a
GZIP console. If it is absent, it falls back to regular output.

Furthermore, two lines of meta information will not be gzipped with the
bulk of the dump, as it is both unnecessary and probably only hurts the
compression.

Change-Id: I8d544eb80290fc3ab620e20507ea2926dd1108e1