extend Mem_unit::make_coherent_to_pou() by size parameter
This function is supposed to be called for a certain data range. The
existing function Mem_unit::make_coherent_to_pou(addr) worked only on
a single cache line.

On ARM, actually call Mmu::flush_cache(). This function does more than
necessary (write back + invalidate D-cache instead of just clean the
D-cache) but we also need to invalidate the I-cache and a convenient
function is currently not available for the ARM implementation.

On MIPS, use the 'synci' instruction synchronizing all caches to make
instruction writes effective.

This function is required for MIPS and ARM but is a NOP for IA32/AMD64.

Change-Id: I7a48a40917fe71cdf8d9daca98c176da2f999a49