ia32/svm: Add missing rw msr permission for MSR_STAR
Fiasco does not currently allow access to the STAR register in the MSR
permission bitmap. So if the guest tries to access the STAR register,
this results in an MSR access error.

The VMSAVE and VMLOAD instructions save and restore the the STAR MSR.
For the related CSTAR, LSTAR and SFMASK register Fiasco already grants
rw permissions. Thus it should be safe to grant the guest rw permissions
for its copy of the STAR MSR.

Change-Id: I98014b5241cbe5c0c5c31f463804ac2a6b13d479