utest_fw: don't call destructors when tests finished
The exit() function calls all kernel destructors. This becomes a problem
if a object used by tests, for example a thread, is waiting for the RCU
grace period during destruction. After the RCU grace is over, the thread
continues its destruction -- which is not possible anymore after the
kernel destroyed itself.

Change-Id: I9989fb90566b0499fb302aab49f32ab180efc62e