Check CTRL_RWP bit after writing to certain registers
The GicV3 spec says:
This field tracks writes to:
• GICD_CTLR[2:0], the Group Enables, for transitions from 1 to 0 only.
• GICD_CTLR[7:4], the ARE bits, E1NWF bit and DS bit.
• GICD_ICENABLER<n>.

Furthermore for the ICENABLER bit it states:
Completion of a write to this register does not guarantee that the
effects of the write are visible throughout the affinity hierarchy. To
ensure an enable has been cleared, software must write to the register
with bits set to 1 to clear the required enables. Software must then
poll GICD_CTLR.RWP until it has the value zero.

Change-Id: I03415af62894fdcf18a2593eda9a8869e7e62818