Arm/GIC-Dist: Consider pin numbers < 32 for set_cpu()
Userland is able to trigger 'Gic_dist::set_cpu(pin)' for pin < 32.
Hence, add a proper check to the GICv3 version because there, according
to the spec, GICD_IROUTER<n> for n=0..31 is reserved. In contrast, this
is not an issue for GICv2 because GICD_ITARGETSR<n> for n=0..31 is RO.

Change-Id: I1030ffe8291d4da5b45e1babd00bc310ddd05100