utest: Fix race between thread_fn and start_thread
start_thread is passed fn by value, so it resides in its stack frame.
It then stores a reference to it in a Thread_args structure, again in
its stack frame. The Thread_args structure is passed to the new thread
and start_thread busy-waits until thread_fn signals that the new thread
is no longer using it. To that end, thread_fn must not use a reference
to the start_thread's stack frame when invoking fn.

Change-Id: Iff34ac063ee0f3ac52bd8dfc215a09af1abaea45