fpu: Allow dynamic selection between different Fpu_state types
This is in preparation of an upcoming change that introduces support
for the ARM Scalable Vector Extension. Because then it is necessary,
depending on whether a thread uses the Scalable Vector Extension,
to maintain a differently structured (larger) FPU state.

The dynamic Fpu_state type selection feature is optional, so for the
case that support for the Scalable Vector Extension is not enabled in
the kernel configuration, there should be no performance degradation
in comparison to the status quo.

Change-Id: I830d47c7f205c39401a65faebb77922a1b589dac