arm64: Add support for ARM Scalable Vector Extension
The ARM Scalable Vector Extension (SVE) extends the SIMD functionality
of the AArch64 execution state with support for larger and flexible
vector lengths.

Because the SVE state can be huge (around ~8500 bytes for the maximum
vector length of 2048 bits), I decided to go for a design where the SVE
state is only allocated/saved/restored for threads that actually make
use of SVE. The main challenge here is that parts of SVE registers are
shared with the regular FP/SIMD registers:
  - The lower 128-bit of the SVE vector registers
  - FPCR
  - FPSR

Per default a thread starts with SVE disabled. If user mode executes an
SVE instruction while SVE is disabled, an SVE trap is triggered in
Fiasco. Thereupon Fiasco allocates an SVE state structure for the
thread, which is used to save/restore SVE register state across context
switches. Then Fiasco enables SVE and resumes the thread.

When restoring an Fpu_state, SVE is enabled/disabled depending on
whether the restored state is an SVE state or an FP/SIMD state. In
fact, this is the only place where the SVE trap flag is toggled. But
since the regular FP/SIMD trap also traps the execution of SVE
instructions and whenever the FPU ownership changes, restore_state() is
called. This is sufficient for both eager and lazy FPU switching.

Current limitations / open issues:
  - While user mode applications running in EL1 can detect the presence
    of SVE via the ID_AA64PFR0_EL1 register, user mode applications
    running in EL0 have no such mechanism. Moreover, even if SVE is
    present in ID_AA64PFR0_EL1, SVE support might be disabled in Fiasco.

  - While user mode applications running in EL1 can configure the
    effective vector length via the ZCR_EL1 register, user mode
    applications running in EL0 have no such mechanism.

This has been tested with QEMU 6.1 and the AEMv-A Base Platform FVP
simulator, with two Linux VMs running on the same CPU executing an SVE
stress test (shipped with the Linux kernel) running in guest user mode.
QEMU supports SVE since version 3.1, SVE support can be enabled with
the '-cpu max' option. To successfully boot a Linux guest I had to
disable pointer authentication ('-cpu max,pauth=off'), which is enabled
per default with the cpu max option.

Change-Id: I85f658c93c2f8a232159263efbb8e9851437cd63