utest: allow custom config tags for 'expected' files
Sometimes it is not sufficient to have dedicated 'expected' file for a
dedicated CONFIG_XARCH setting and/or a dedicated CONFIG_XARCH-BITS
setting. Also allow to specify

    expected_foo-config = bar

to define a custom config tag 'bar' which is then used to find the
proper 'expected' file. For example, the above setting makes the build
system look for 'expected_foo-bar' for 'test_foo'  before it checks for
'expected_foo'.

Change-Id: I724ec2ffa5a2bbad430d46e3f7e9ad73a8ead06b