utest: provide Utest::kmem_create_clear()
This function does the same as Utest::kmem_create() but in addition the
memory for the object storage is initialized before the constructor is
called. This is required for objects which would normally be stored on
the BSS (which is initialized with zero at kernel startup).

Also change Utest::kmem_create() to always initialize object memory to
0x55 before calling the constructor to trigger bugs due to uninitialized
object members.

Change-Id: I2ea9e2bae3bcf55e553407dcc80227f61f76b2bb