ARM: setup HSTR_EL2 register for ext vCPU
We have to make sure that the HSTR register is setup correctly when we
enable the extended vCPU feature on ARM. Otherwise, the VMM might not be
able to access system registers correctly. The HSTR used to be set up
correctly after a context switch to a thread with extended vCPU enabled
but not initially.

Change-Id: Ib2ddd7fc9fe1a37832eca16d9deef5fab1f209ac