drivers: Put FAILED among console states
The FAILED bit is used together with _state and not as a console
attribute. Move accordingly.

This change also moves the printing of the possible failed state from
str_attr(), where it would never be printed, to str_state(). str_state()
now prints the usual enable info unless the console is failed, in which
case it only says it failed.

Change-Id: I7f3836369827f68b92eb215ed65156b89ae991f6