Ipc_gate: properly handle "label" without receive phase
When a thread C sends and IPC to an IPC gate, the kernel automatically
modifies the label using the value which was specified when a thread S
was bound to that gate.

If the IPC contains a receive operation, thread S bound to the gate
defines the transferred label. Otherwise the label must not be modified.
Change the IPC path to consider this.

Fixes ticket #CD-465.

Change-Id: I561bd5cf145cf25582350fee9e359a3470978eae