Arm64: Remove CPUECTL handling
Enabling SMP is not necessary anymore on any recent cores, because it's
always enabled. Further, the MSR/MRS allocation for CPUECTL has changed
on more recent cores, making the current one an invalid one.
Even more, booting via PSCI and other similar boot loaders, it is the
task of the boot loader to enable SMP. Thus remove this whole
functionality as it is not needed on Armv8/64 systems.

Change-Id: I8cd643f53b2a761624201ba1e8e2ef5e9c4683a2