Add Mem::[read64/write64]_consistent
This functions allow consistent reads and writes to 64bit values.
This means on 32bit systems write are done LSB first with a barrier,
and reads are done MSB first with a loop comparing the MSB value after
reading the LSB. This is useful when updates to some 64bit counter shall
be seen atomically by a reader.

Change-Id: Ic0ad2990ec9dfec402c243e81309cc6b4145c126