Unify copies of Mem_space::switchin_context
Although the method differs slightly between the architectures, this is
probably more due to historical development rather than a conscious
decision.

Some architectures use `if(dir() == Kmem::kdir)` to determine if the
space switched to is the kernel space, others use
`if (this == kernel_space())`. Both should be equivalent, because the
kernel space always uses the Kmem::kdir as its page directory and is
also the only space doing so.

Furthermore the ppc32 variant contains a broken optimization, which is
disabled on ARM and ia32. This is probably an oversight, or maybe it's
because the ppc32 port doesn't have SMP support. In both cases,
maintaining the variation appears to be unnecessary.

Change-Id: Ifb9533dde9f9df8e033bcd7eb9560829b6e3c70c