arm: Add support for platforms with multiple ITSs
There can be multiple ITSs in the system, which might each be
responsible for a different subset of devices. Thus now, the assignment
of an LPI to an ITS is not decided until user space binds the
corresponding MSI to a device. Therefore, I had to move the management
of LPIs from the Gic_its class to the Gic_msi class (MSI interrupt
controller).

Change-Id: I774c216e5e50499f60964875159a3d328bb7565b