arm: fix double free on CONFIG_ARM_PT48
When a task is destroyed, its page tables need to be freed. Depending
on the actual platform, there might be pages above Mem_layout::User_max
that either belong to the kernel or to the virtually mapped caps array.
These page directories must thus be freed too.

In case of arm64 && arm_pt48, the "Virt_addr(Mem_layout::User_max + 1)"
expression will yield an address that will exactly overflow the virtual
address range of the page tables:

  User_max(=0x0000ffffffffffff) + 1 == 1ULL << 48

Because the page table virtual address space has 48 bits, the upper
bits will be masked by the page table walk code and iteration starts
effectively from Virt_addr(0). This leads to a double free in destroy()
because the same page tables are iterated twice.

Introduce a constant in the page tables code to have a consistent
definition of the maximum user space virtual address. Also use this
constant in the other architectures for the sake of consistency.

Change-Id: If119e4e8f10be5892a597af64c31f1f8c8a535d7