arm: add infrastructure for runtime code patching
Provide a low-level macro ASM_ALTERNATIVE_ENTRY() on arm that selects
between two variants of inline assembly depending on the return value
of a function. This function is called early during boot and the
respective variant of the assembly is used.

Building upon this low-level facility a high-level "static boolean key"
is provided. Such a key is evaluated once at boot time. It can be used
in boolean contexts like any other variable but the actual object code
will have static branches that are patched depending on the functions
return value.

Change-Id: I294c58e87b599f3422586ca27d67340fe39254d7