Use the typename keyword in bitfield macros
Without the use of the typename keyword in the return types of the
accessor methods of the bitfield members, the bitfield members cannot
be used within template classes. In such cases, the compiler reports
that the return type name is a dependent scope and without the typename
keyword the name lookup is ambiguous.

This is purely syntactical change and it does not change any
functionality.

Change-Id: I751b5e340f42a82a29b459711de714fba6599130