arm: support hardware with ITS page sizes != 4k
An ITS implementation is not required to implement all page sizes but
might have a read-only Page_Size field in GITS_BASER<n> with its single
supported size. We always try to configure 4k pages to keep the
possible alignment overhead in the buddy allocator low. But if the
Page_Size field is read back with a different value we use the hardware
enforced page size.

Change-Id: Id1a9f99692394448bd5a8bcf71f81f050fe72d92