context: properly synchronize on _running_under_lock
The _running_under_lock variable acts as a spinlock to synchronize
helping attempts of other Contexts when they want to acquire a lock
presumably held by this Context. If the lock was acquired we need to
make sure that no speculation happened on values that are access after
this point. Likewise, if the lock is released, any modifications made
by the caller must be observable to other CPUs before the lock is
visible as being unlocked.

Fixes ticket #CD-280.

Change-Id: I06df2fab027e6083e93095ca1c697a38fd98c8c1