utest_fw: introduce Utest_fw::chk()
The Utest_fw::chk() function shall be used to verify conditions in unit
test setup / cleanup code which are not relevant for the actual test.

If the parameter `result` is false, the function will print an error
message and abort the test. In contrast to the UTEST_xyz macros,
Utest_fw::chk() expressions are not logged on (expected) success.

  Utest_fw::chk(false, "This condition should be true.");

In case of failure, the following output is generated:

  KUT not ok 1 Mapdb::test_foo_bar
  KUT # Test-uuid: 8f94c191-6bee-4b66-aa94-ff5dd967865b

  KUT # Failure in test setup: This condition should be true.

Change-Id: I3841269f17d3c1f697f00350f79c68dc4458c6e6