Fix reference leak in thread kill
The Thread::kill() method can be called concurrently. One possible path
is the deletion of the thread object through capability references.
Another possible invocation is the threads page fault code if
the threads pager is invalid. This might happen on distinct CPUs.

Fix the race by taking the additional reference to the thread only once
in the kill path. The reference can be safely taken in the DRQ handler
in the cross-cpu case because the original caller is supposed to have a
reference and is waiting for the DRQ to finish. If the thread is dying
or even already dead, the concurrent invocation must not do anything.

Change-Id: I593e34738ad276d3bff1abd9df9c3f23579019af