S_list_tail: fix move ctor / assignment op from empty list
Properly handle the '_tail' member if a new list is constructed from an
empty S_list_tail or if an empty S_list_tail is assigned.

In that case, the '_tail' member has to point to the address of the '_f'
member of this object, not to the address of '_f' of the other object.

Fixes ticket #CD-569.

Change-Id: I5050a9ab0b3d925dfad5666052f1334899793eab