arm: don't configure PPIs via distributor on GICv3
Setting the interrupt configuration (level vs. edge triggered) for a
PPI through the distributor is only working on GICv2. It is
theoretically working on GICv3 when ARE is not used but we do in fact
enable ARE. Hence we must not access GICD_ICFGR0 and GICD_ICFGR1 on
GICv3.

To make the per-CPU semantic explicit a set_mode_percpu() method is
added, just like the already existing per-CPU mask/unmask methods. On
GICv3 this will then write to the respective redistributor registers.
The only call site (timer_tick-single-vector.cpp) needs to be adapted
accordingly to use the new facilities.

This fixes a regression introduced by bcd092a07deb. Before this commit
PPIs could not be configured and no undefined behaviour was invoked.
Depending on the hardware it could prevent the timer interrupt from
working reliably.

Change-Id: I9303358dc7383bd0c40a639c152556aa66a9d111