arm: fix page attributes on cpu_virt
When running on arm with lpae page table format and cpu_virt the user
space is running with a stage-2 page table. This page table format
stores the memory attributes directly in the page table. To get normal
uncached memory MemAttr[3:0] must be 0b0101, thus 0x14 must be stored
in the PTE.

Change-Id: I9ab71d03b5413a726a8c721d84f8f82ab30c3555