Fix triggered assertion in Context:drq()
There is a race condition between binding an IRQ and triggering it that
causes an assertion in Context::drq() to be triggered. For details, see
ticket #CD-679.

This commit fixes that by removing the resetting of `_queued = 0` if
there was no bound thread. To compensate for that, we now check the
thread is valid before queueing when the IRQ is triggered. This way,
`_queue` does not increase at all when no thread is bound, so we also
don't have to reset it.

To prevent an interrupt storm, we now mask_and_ack() instead of just
ack() for level-triggered IRQs. This would break a user that expects an
IRQ that was just attached to be unmasked, instead of manually unmasking
it. But that previous behaviour is considered an implementation detail
and users doing this broken.

Change-Id: Icd29179d2d241d940ec2c0eaa5be9977110ac2d8