Unify implementations of Mem_space::virt_to_phys_s0
The Mem_space::virt_to_phys_s0() method was implemented for each
architecture as a copy of the respective Mem_space::virt_to_phys()
method.

With this change, there is now only one architecture-unspecific
implementation of this method, which simply calls virt_to_phys().

In addition, a note has been added to the method's documentation on how
the Sigma0 special case is applied.

Change-Id: I73f6fc1e53329d1fbf66baaff1c58e074de8b650