arm: Refactor inconsistently used Cpu::init_mmu method
The method Cpu::init_mmu was used to set up the exception vector
in a non-virtualization-enabled Fiasco. On ARM 32-bit this includes
mapping the exception vector page to a specific address, which was the
origin of the method's name. However, on ARM 64-bit no page needs to be
mapped, there only the address of the exception vector is written to the
VBAR_EL1 register.

So a more accurate name would be Cpu::init_exception_vector, which
however turns out to be inadequate as well, because in a
virtualization-enabled Fiasco the exception vector is already set up as
part of the Cpu::init_hyp_mode method.

In addition, for the upcoming support of 16-bit ASIDs, we need a method
to check during CPU initialization whether 16-bit ASIDs are supported.

Thus the following refactorings:
 - Rename Cpu::init_mmu into Cpu::init_supervisor_mode.
 - Unify initialization order between the boot and secondary CPUs,
   i.e. since Cpu::init_supervisor_mode() is now invoked by Cpu::init(),
   during startup Cpu::init() must be invoked before Pic::init() and
   Thread::init_per_cpu(). For secondary CPUs it was already done in
   that order, but not for the boot CPU.

Change-Id: Ib65c5c6fc81cd73b222e6bfd431d92a6cdea638c