Fix "Guest IA32_SYSENTER_EIP" VMCS encoding
According to Intel 64 and IA-32 Architectures Software Developer's
Manual, Section B.4.3, the natual-width guest-state field of the VMCS
with the key "Guest IA32_SYSENTER_EIP" has the encoding 0x6826, not
0x682e.

This change fixes the typo in the constant. Since this is a fix, not
a change of the intended semantics, we keep the Vcpu_arch_version
unchanged.

The order of declarations of two 64-bit control constants is also
corrected.

Change-Id: I46dff677e8222588843f2fcfd3137d701cf8cb2d