kip: Improve documentation of system clock reading code
The "might or might not be synchronized with the KIP clock" statement
also applies to the code for reading the system clock with microsecond
resolution.

Furthermore, the term accuracy has been replaced by resolution because,
depending on the configuration and platform, the accuracy/precision of
the clock may well be lower than the resolution of clock value the
reading code provides.

Change-Id: Id97a794b917067f9f17cc149d005a36a271b31a5