VMX: Avoid kernel crash on VMX resume failure
The error path in do_resume_vcpu() after a failed resume_vm_vmx() call
needs to properly restore the host state before the failure can be
signalled to the caller. Otherwise the kernel might crash after
switching to a different context due to unexpected flags in some
control/machine specific registers. One such scenario is the XRSTOR
instruction causing a general protection fault due to the fact that it
encounters a bit set in XSTATE_BV that is not set in the XCR0 value for
the guest.

See ticket #CD-825.

Change-Id: I6362e52e2a4c964f0cc2a24bd28eca7a7ee0008e