arm: prevent unpredictable behaviour on Armv7
If a thread is switched into extended vCPU state the user space has
control over both HCR.TGE and the CPSR.M of vCPU user mode. The Arm
architecture defines that if HCR.TGE is 1 then executing in a PL1 mode
is unpredictable behaviour. On Armv8 the same condition will trigger an
architecturally defined "invalid exception return" but there is no code
to handle it either in Fiasco.

Lets be on the safe side and sanitize the vCPU state. If HCR.TGE is 1
force PL0 (user) execution mode.

See ticket #CD-853.

Change-Id: I7f915bb35359d0f0895198fc88cd07edb223ae39
Requires-Change-Id: I938ba50848a6f81b66cca2242ed744bd1aee1b9e