arm: always treat physical memory as non-contiguous
This unifies the kernel heap handling on all arm platforms by requiring
that the heap is always mapped through a dedicated Pmem area. It
removes the implicit assumption that there is a fixed
virtual-to-physical offset of the whole heap. Because the heap is now
at a dedicated virtual address in the kernel, any code that relied on
the 1:1 mapping assumption of cpu_virt configurations must use the
proper pmem conversion functions.

Using a dedicated virtual address region for the kernel heap is the
first step to make the kernel independent of its physical placement in
RAM.

Change-Id: I833a891d45ba639d0cce2fea2a4c720f9e15a60b