Irq_sender: don't allow to bind any thread after destroy
After an Irq_sender object is destroyed, no thread must be bound to the
Irq_sender object. This is achieved by trying to acquire the existence
lock before a thread can be bound to the Irq_sender object.

Also, during destroy, call Irq::destroy() before detaching from a bound
thread to perform the release operation with acquired existence lock.

The existence lock is acquired when a kernel object was destroyed and
the lock is not released until the object is finally deleted.

See also ticket #CD-809.

Change-Id: Id188f911221cc57a04f57e44696451516ce2589e