x86: Rework MMIO in-kernel remapping
Remove the Device_map class which was only used once, and could only be
used for a single superpage, and replace by mmio_remap as used on the
Arm architecture.

With this, simplify the APIC and ACPI codes to use mmio_remap, and
remove some special casing of x86's Lapic page.

On arm, also remove the Device_map adaption.

Change-Id: Ifc8f27747c78d2b390cae88aa0d68699526626d3