ia32/svm: use unique kernel address with kernel_vmcb()
Svm::kernel_vmcb() has the precondition on the caller provided VMCB
address to be unique across all address spaces. This precondition is
violated in configurations with Config::Access_user_mem_direct or
Config::Must_access_user_mem_direct set, because the calling code
operates on the user-space address of the VMCB. This address is not
unique and can be used in multiple address spaces.

To meet this precondition we always use the kernel address of this
KU memory when calling into Svm::kernel_vmcb().

Refactor acquiring the VMCB pointer into a function.

Another issue related to the clean bits is described in ticket #CD-920.

Change-Id: I0a34370b4ee375b46fc72802dce2dcedbc37e0dd