Arm: gic_mem: Add option to force non-shareable/cacheable
On some platforms (e.g. Qualcomm SM8150) the GIC seems to accept
cachability/shareability memory attributes even though it does not
support them (because it is not cache-coherent). This breaks the
discovery mechanism in Fiasco and causes trouble when commands are
parsed by the GIC that have not been flushed from the cache yet.

Add a workaround for this using a ARM_GIC_NOT_COHERENT Kconfig option
that can be used to force non-shareable/cacheable without using the
discovery mechanism.

Change-Id: Iec37caae5ff6b2991d2f3dcc8d298796c419df05