mark arm_fast_exit() as not returning
The function arm_fast_exit() will not return under any circumstances
because the implementations unconditionally jump to an address passed as
'pc' parameter.

As the compiler cannot deduce this fact from the Assembler code, use the
__builtin_unreachable() built-in function and add "noreturn" to the
function declaration so the callers know as well.

This removes some dead code from the binary.

Change-Id: I9edbc0ba0446a764a7a0a5b01afb33d9b1f15b4b