map_util: Fix use-after-free in Auto_tlb_flush during map/unmap
The Auto_tlb_flush class keeps ephemeral references to Mem_spaces in the
form of raw pointers. Once all map operations have been completed and
the corresponding mapdb frame locks have been released, the flushes
stored in Auto_tlb_flush are performed. This can lead to the dereference
of a pointer to a Mem_space that has been deleted in the meantime, i.e.
the memory pointed to might now be used for something else.

The reason why a Mem_space may have been deleted in the meantime is that
interrupts are open during map/unmap and its existence lock is not held,
so ephemeral references to it can vanish at any time (RCU grace period).

Therefore, to ensure that a Mem_space is not deleted, Auto_tlb_flush
must increment the space's reference counter when storing the reference.

See ticket #CD-972.

Change-Id: I348a90752b00e6f884536f27f496f9541bbc5bd0
Requires-Change-Id: I48f5193e0eae12d13ef0788a60308c63d09405e0