semaphore: Reset thread partner consistently
So far the thread partner was only reset in case of a timeout or cancel.
Get rid of this inconsistency, i.e. reset the partner in all cases.

Note that this effectively causes no change in behaviour, because
resetting the partner is not strictly necessary, as the partner is only
taken into account if the thread has set the Thread_receive_wait or
Thread_receive_in_progress state flag. And wherever one of these flags
is set, the partner is also updated beforehand, i.e. the old value is
overwritten. The IPC code paths, for example, never explicitly reset
the partner.

Therefore, a possible consideration for the future would be to
completely refrain from resetting the partner in semaphores as well.

Change-Id: Ie006979fed18b1d950421e24aa61a69ec1fdfdfc