arm: always use generic timer virtual counter
Qemu is buggy when checking permissions to access the physical counter
from EL0. To work around this always use the virtual counter,
regardless of the counter that is used by the kernel. In case of
cpu_virt we just make sure that CNTVOFF is 0 whenever regular user
space is executed.

As a nice side effect we do not need to update CNTHCTL on context
switches any more. This speeds up switches from vCPU kernel to user
mode and context switches to/from VMs.

Change-Id: I94d99ae50c8fec8c3509af41a63cd988541d60ce