vkey: Remove unused Vkey::add_char() function
The Vkey::add_char() function is unused since
change I0485229ea91bbd3c36426fbaee916a70fe95a281.

The underlying Vkey::add() function is not thread-safe, so it would not
be advisable anyway to pass input to Vkey from anywhere else than via
the call to Vkey::check() in the UART IRQ handler or if no UART IRQ is
present in the timer tick handler.

Change-Id: I6aaaa68aa967f205f80cee59a73b562d32a7ddc5