vkey: Ensure memory order for cross CPU consumer and producer
In case the Vkey producer, i.e. Vkey::add(), and consumer, i.e.
Vkey::get(), run on different CPUs, it is not guaranteed that if the
consumer sees the updated vkey_head index it also sees the
corresponding character in vkey_buffer. The producer CPU might have
reordered the two writes and/or the consumer CPU might have reordered
the two reads.

A similar problem exists in the opposite direction, there the consumer
CPU must ensure that it updates the tail index only after reading the
corresponding character from the vkey_buffer.

Add barriers to ensure that the consumer sees the write to vkey_buffer
before it sees the updated vkey_head index, and the producer sees the
updated vkey_tail index only after the consumer read the corresponding
character from the vkey_buffer.

See also ticket #CD-1033.

Change-Id: Ia2f051d27aeda0b7e0d0172f697c5d6f812ff593