arm64: resolve cyclic dependency
To calculate the maximum supported user space address and the desired
UTCB address of sigma0 and the root task, we have to consider the page
table configuration and the CPU feature bits. To prevent a cyclic
dependency, move the calculation of the UTCB address to the
Kernel_thread class where it is needed. The maximum supported user
space address is provided by the Mem_space class which deals with the
user memory space. By default, the values from Mem_layout are taken.

See ticket #CD-1038.

Change-Id: Id4681bae53cc3e9f4dff5e13f876634598d99a62