ia32: Replace magic values with Xstate constants
The Fpu module used a magic value for the defined Xstate bits, although
a constant was already defined for this value, but in the Vm module,
thus not accessible for the Fpu module.

Therefore, move the Xstate contants into the Cpu module which is
accessible from both Vm and Fpu.

Change-Id: I1f46cc1b892a9cb40a9cdcfab9e13bf445416370