Cleanup: remove mini_defs.h
The mini_defs.h header is non-standard (within the context of libc), its
only definition of MARK_DEPRECATED is currently not used anywhere and
the branching in the definition is also no longer necessary. In
addition, if we need the deprecated attribute, we can use it directly
(as we do with other attributes).

Therefore it is safe to remove the header.

Change-Id: I9cd34753f5505c227ee59d37d311cec90b78af8b