thread-ipc: consider timeout when activating IPC partner
This is another attempt to properly consider the receive timeout when
switching to a partner after a successful send phase before the receive
phase is started.

We need to set up the timeout before activating the partner because the
timeout might hit too late if the partner runs on a large timeslice.

While at it document the conditions that have to be met for directly
switching to the partner (timeslice donation).

See ticket #CD-434.

Change-Id: Ibfcc037edebc67aa47e2026f77ac813acba34838