kobject_mapdb: keep ref_cnt consistent even on object deletion
When an object capability is unmapped without requesting immediate
deletion, `ref_cnt` was decremented twice when unmapping pushed
`ref_cnt` below 1: Once within `flush()` and again in
`invalidate_mappings()` because the entry was not removed from the
entry list in `flush()` in this case; presumably for optimization.

This change restructures the code while keeping the optimization but
only decrementing `ref_cnt` in `flush()` if `invalidate_mappings()` is
not called.

Note that the previous behavior was not an issue because the `ref_cnt`
of an object is irrelevant after all mappings for the object are gone.
However, an inconsistent `ref_cnt` might be confusing when debugging.

Change-Id: I92c600c87693b2c901dfe5cc18d3d12f006333f2