thread-ipc: adapt comment for failed send operation
The original comment says that we don't switch to the IPC partner and
skip the receive phase if the send phase failed.

The original code in change 1e3fc84b9d361af964b321f0f55086342acbc9ee set
have_receive to false, most likely to explicitly skip the receive phase.
However, this variable is overwritten a few lines below and therefore
this statement did never have any effect. Consequently it was removed in
If275b25e50f5cfe22030cfa38fe466050951155b.

If the sender check returned OK (partner either in open or closed wait)
and Thread::transfer_msg() returned an error (see the default case in
the switch statement a few lines above), activate_partner was set in
the usual case (the IPC should not be sent to the current thread). In
that case, we activate the partner (mandatory to get the partner running
again) but also might switch to the partner.

As the Thread_receive_wait flag is not set for this case, the receive
phase is skipped.

Adapt the corresponding comment properly.

Change-Id: If726ff82eb54e043b929c33e636525c01592d67e