Make the x86 descriptors code type-safe and fix usage
This change makes sure that the logical meaning of specific bits in the
code that configures x86 descriptors cannot be easily confused thanks
to the use of strictly-typed enums. Several impossible combinations of
bits are also guarded by assertions.

Specifically, the incorrectly pre-shifted value of the 4K-granularity
bit in Context::arch_setup_utcb_ptr() and the incorrect use of the
accessed bit in Cpu::init_tss_dbf() are fixed.

Documentation comments are added.

Change-Id: I94200413307c3c3c7e005eac5e1ceaa2fb780af5