provide dummy operator delete(void *, std::align_val_t) for C++17
According to C++17 the compiler will generate code to use

  operator delete(void *, std::align_val_t)

rather than

  operator delete(void *)

for deleting objects if the alignment of the object exceeds the value of
__STDCPP_DEFAULT_NEW_ALIGNMENT__. So far this was only seen on SPARC
with __STDCPP_DEFAULT_NEW_ALIGNMENT__==8 and alignof(Context)==16.

As we don't free runtime objects this way, the default implementation of
the delete operator is a dummy triggering panic(). However, we shall
provide a dummy for the C++17 delete operator to be on the safe side.

Even Clang-17 doesn't provide std::align_val_t, so this is gcc-only.

Change-Id: I4920b277906677acf5ac31a7436dbf8f0c131895