Move kdir from Mem_layout to Kmem
In mem_layout.h it is not possible to include paging.h, as that would
create a circular dependency. Therefore a forward declaration of Kpdir
was used. This then had the disadvantage that users of kdir in addition
had to include paging.h themselves.

Besides, kdir seemed out of place in Mem_layout anyway, which is
otherwise really just a collection of constants describing the layout of
the address space, as also the name suggests.

See ticket #CD-1101.

Change-Id: I7ce94ac71cb1b335a60d473f8840d936b904bec5