boot/{ia32,amd64}, jdb_bp: convert a few macros to inline functions
This fixes a few pedantic warnings about "braced-groups within
expressions".

The Assembler statements for reading debug registers are tricky: The
functions have to be always inline because the passed register number
must be constant and known at compile time. The '%c1' Assembler
constraint delivers the register number as raw value to the Assembler.

Change-Id: I7585a59f3dac294b9935a96faa35c1c49de1efc9