thread-ipc: handle missing case for transferring FPU state
Following scenario: The FPU state is transferred via IPC (Transfer_fpu
flag set in message tag and the receiving UTCB has the Inherit_fpu flag
set) and both, sender and receiver, are running on a different CPU. The
receiver gets ready to receive when the sender is already part of the
receiver`s sender list.

In that case, the receiver calls Thread::ipc_send_msg() on the receiver
CPU in the context of the waiting sender. The Thread::transfer_fpu()
function enforces that in case sender and receiver are on a different
CPU, both threads must not be the current FPU owner on their CPU.

This is achieved for the sender in Thread::remote_handshake_receiver().
Also, in case the receiver was initially ready to receive the IPC, the
receiver would have spilled the FPU state in Thread::remote_ipc_send().

The same has to be done in Thread::ipc_send_msg() for the receiver (but
only if sender and receiver are executed on different CPUs), otherwise
an assertion in Thread::transfer_fpu() may hit.

This is actually a bug fix for this case -- without this fix the kernel
would just not transfer the FPU state in that case (and trigger the
mentioned assertion in debug builds).

See ticket #CD-1141.

Change-Id: I0852bbda20ea7626c15e31ffa87756a780584627