ipc_gate: Fix handling of canceled threads waiting on unbound gate
Threads that try so send to an unbound IPC gate block until a thread is
bound to the gate, the send timeout expires or the thread's IPC
operation is canceled, e.g. via ex_regs().

The cancel case was not considered by Ipc_gate::block(). When a thread
that was blocking on an IPC gate was canceled, it remained in the queue
of the IPC gate and an incorrect error code was returned.

In addition, the Thread_timeout and Thread_cancel flags were not
cleared, as they are not part of Thread_ipc_mask.

See ticket #CD-1148.

Change-Id: I046a2f63ecf09704f4b1309b454b76c5bf6af69a