tb_entry_output: refine printing of msgtag and words during IPC
For tracebuffer entries logging IPC and IPC results, show one or two
dwords only if the message tag announces at least the required number
of dwords. If more than 2 dwords are part of the message, show ",...".

On the IPC result side, interpret the entire message tag if the error
flag is not set. Show the dwords only if there was actually a receive
phase.

Change-Id: I29b15d82cef8417ef35904990c02fdc85465befb