fix a few edge cases in interpreting IPC timeouts
For non-zero and non-infinite timeouts, properly consider the case when
the timeout has already expired. In particular consider that an absolute
timeout has expired before starting the IPC send phase and when blocking
on an unbound IPC gate.

Also, consistently use 'tval' rather than 't' for variables containing a
time as 't' is often used for threads.

See ticket #CD-1140.

Change-Id: I79b78875fcf7dfc9fb2657c96bd01f7689491ce6