Replace extended vCPU offsets by constants
The extended vCPU infos are at offset 0x200 into the vCPU structure.
The extended vCPU state is at offset 0x400. Instead of using these
magic values everywhere, replace them by constants.

The vCPU state size is defined as constexpr function because the
PAGE_SIZE definition is added by the architectures which will always be
after the common definitions. Fortunately it is possible to use a
constant in an inline function that is not yet defined.

Change-Id: Iaaf44971a495961068ba7d66d193a5874116bd1f