arm: fix PMU event counters in virt-build
This change introduces a check for the processor feature FEAT_HPMN0.

If this feature is available, we retain the current behavior and reserve
all PMU event counters for use in EL2.
Otherwise, we assign at least 1 PMU event counter to be accessible from
EL1. This is not a problem because we trap PMU register access anyway by
setting MDCR_EL2/HDCR.TPM to 1.

See ticket #CD-1163.

Change-Id: I4e70f6bd56ad3ddb867926a07cd26ba563a6d6a6