Icu_h<>::icu_invoke: check for number of passed mwords
The callers of this function don't invoke Ko::check_basics() so check
for the minimum required number of passed mwords here.

We don't need to check for the prototype (this function is only called
on tag.proto() == L4_msg_tag::Label_irq) and not for rights.

While at it, use "tag" instead of "t" as variable name for the message
tag for consistency reasons with other kernel code.

See ticket #CD-1102.

Change-Id: I010e91116a14e8dd575f8564c8ccc1504c65192d
Requires-Change-Id: I4c7f43a477d26c87f8b26e170de36a713bae9e38