delayloop: improve reliability on certain platforms
Delay::delay() is used wait for a certain period of time when there is
no timer tick is available -- in JDB. Delay::measure() executes a loop
and determines the number of turns with reading from the KIP clock until
the clock advances by one millisecond.

It is important that Delay::measure() executes code similar to
Delay::delay() to have a reasonable accuracy. Add barriers to prevent
the compiler from using conditional return statements within the
measurement loop.

For example on ARM32, without this change, the code actually waits for
less than 1/4th comparing to the intended period.

Change-Id: Ib7e2fb4eec5dc64a4635075b2119e69e20e2cb48