gic: fix init crash on platforms without firmware
On GICv3, the system register access must be enabled explicitly. This
is done when enabling the CPU interface. However, the generic GIC setup
code briefly disables the CPU interface, only to enable it later in the
GIC version specific constructor. This won't work if the GIC CPU
interface register access was not yet enabled by a firmware.

As we enable the CPU interface anyway at construction time and because
interrupts are still disabled at the time, just remove the early CPU
interface disabling.

Change-Id: Ie2b748b9ba817cc39bfbbe56fd8d5a86d8005261