arm: bootstrap: initialize HCR_EL2 before enabling paging
The value of HCR_EL2 is "architecturally unknown" after reset. The
hardware or a previous bootloader might have initialized them in a way
that causes trouble for Fiasco. HCR_EL2 is initialized in
Cpu::init_hyp_mode_common() but this happens quite late in the Fiasco
start process.

A problematic configuration is HCR_EL2.{E2H,TGE} with values {1,1}.
This enables the Virtualization Host Extensions (VHE) in which several
EL1/EL2 registers are redefined. In particular, the normal EL2
translation regime is replaced by the EL2&0 translation regime that
behaves more similar to the standard configuration of Fiasco in EL1
(with virtualization disabled). Having these bits set interferes with
the Fiasco startup process because paging behaves incorrectly.

Avoid this by initializing HCR_EL2 with a known value before enabling
paging. This is already done in Bootstrap::leave_el3() so we can just
move the existing code to fix the issue.

Change-Id: I4c1f5abb81526c8b157644b512cebd17a39f1cc7