Switch_lock::wait_free: must not be called if current() is owner
Panic if the current thread owns the Switch_lock and attempts to destroy
it. The old code handling this case by clearing the lock was dangerous:
Assuming that the lock owner acquired the lock using a lock guard, that
lock guard will eventually try to release the lock, triggering a nullptr
de-reference in Switch_lock::clear_no_switch_dirty().

See ticket #CD-1280.

Change-Id: I169da2e676f93751afd0b9d26720277afac3734c