ia32: detect invariant/unified TSC
Detect this features by reading CPUID.80000007H:EDX[8].

This feature guarantees that the TSC runs at a constant rate in all ACPI
states (P/C/T) and the TSC is synchronized across cores and sockets but
not across NUMA nodes or processor "cabinets".

The initial synchronization is already ensured by setting the TSC_ADJUST
MSR to 0 on the boot CPU and on all application CPUs.

Change-Id: I18eb2d9a295c1388eabbaa12b3cb2a7413f26b4b