ia32/cpu: make some fns FIASCO_CONST and add 'inline' asm qualifier
Add `__attribute__((const))` to suitable Cpu functions which means that
passing the same argument twice to the function will return the same
result in both cases. If such a function accesses a time stamp scaler
variable, the scaler does not change between the two invocations by
definition. This is true because the scalers are never changed after
initialization.

Also add the 'inline' gcc asm qualifier to suitable code requesting the
compiler to assume that the generate code is very short. The default
mechanism for determining the code size of generated Assembler code is
quite conservative.

The volatile asm qualifier is not required for functions which only use
input parameters to evaluate the result. Remove it in such cases.

Change-Id: I1d34851540ec7f340723b8ce348c88dbad060aa4