fix gcc-14 warnings about uninitialized quota in delete operator
These changes fix warnings triggered by gcc-14 because the compiler
assumes that the _quota field is not initialized (anymore) in the
delete operator.

The compiler may assume that the object is not valid anymore when the
delete operator is called, hence, the compiler warns if object members
are accessed from there. This change prevents the warning by telling the
compiler that the to-be-deleted object has some defined content.

Even without this change, most of these warnings vanish on setting
CONFIG_OPTIMIZE_FOR_SIZE=y. Maybe some heuristics is disabled when
less code is inlined by the compiler.

The orioginal commit used asm stuff, here we use cxx::launder /
std::launder to make the deal... also some additional places where this
is needed.

See also ticket #CD-1323.

Change-Id: If2f663de6a3141a8958204e51aadff2319822b22