libc: Remove buffered read logic
All use cases initialized structures with a limit of 0 anyways (meaning
that all data has already been read and is present). So remove some of
the unneeded complexity.

Since for string buffers shend was always set to the value of rend and
rend was always set to -1, __uflow and __shgetc would never be called
unless memory would wrap while processing the string. Thus this code
would never execute.

Change-Id: I0c0170d8db08508e350f0f87021cb3c9b6d76334