jdb: enhance {peek,poke}_task() to avoid crossing page boundaries
Both functions use Jdb::access_mem_task() delivering the virtual address
of a mapping. This mapping is usually a single page without a guarantee
that the following mapping points to the following physical page. Hence
we must not cross the page boundary when accessing memory at this
address.

Also change the size type to size_t.

Change-Id: I362a56a6b3b9f6021440f0d2f2840713f0513f37