jdb_tcb: another cosmetic fix for TCB dump
The function for displaying (on screen) or dumping (for reports) the TCB
calls architecture-dependent functions to print register information
corresponding to the current thread state Jdb_tcb::info_thread_state()
and to print return frame registers Jdb_tcb::print_entry_frame_regs().

Not all architectures implement both functions, so make sure that these
functions (if implemented) finish with printing a CR and don't do this
in the calling function.

This does not change the screen layout because the following stack dump
uses absolute screen coordinates.

This change is basically ensures that for dumping a report there is a
single empty line after printing the register information before showing
the disassembled instructions at the instruction pointer.

Change-Id: If523da3c3fa63209a39f59aeed046c896160b543