gic: ignore polarity when setting line mode
The GIC can only distinguish between level and edge triggered
interrupts. The polarity can neither be configured nor is it defined
architecturally.

Specifically, PCI legacy interrupts are defined as low-level triggered
interrupts. As io will forward this mode to the system ICU (the GIC in
this case), it is rejected without this change, even though the mode
was correct from a system point of view.

Change-Id: I58bbc666ffac735d4069f3de4bb5cbde5f9d00c8