Make sure the IRQ vector stubs are properly sized and aligned
On AMD64 with CONFIG_KERNEL_NX=n, the actual size of the individual IRQ
vector stubs varies depending on the variant of the jmp instruction
used by the assembler (i.e. due to the distance to the jump target).
This unchecked caused the misalignment of a few last stubs and the
result was a kernel crash due to incorrect stub code being executed
when an IRQ in this range was bound and triggered.

We fix this by making sure that the instruction bundle of each IRQ stub
is aligned separately and we also check the size of the whole sequence
of stubs.

While the code was correct on AMD64 with CONFIG_KERNEL_NX=y and on
IA-32, we employ the same checks for consistency. On IA-32, this
enlarges each IRQ stub to 16 bytes (which should be of no issue).

See ticket #CD-1451.

Change-Id: I9fd716531851af02734673516ae6f6451b43cb34