jdb/objects: improve kernel object list view
Print the name of a kernel object (if available) highlighted to make it
easier to recognize. Also don't align object name and object type name
in non-dense mode to a fixed width because this clutters the view.

Aligning the object type name to a fixed width was broken anyway since
printf does no longer filter-out ANSI EXCape sequences when determining
the number of printable characters of strings.

Change-Id: I6260210887ecd916a8b6f58c5e7c871b33411ae2