Remove unnecessary Address_type enum
Since UX was removed, the Address_type enum is only used in
Jdb_entry_frame and Jdb_bt. Unfortunately there is no
architecture-agnostic source file for Jdb_entry_frame where the enum
could be moved to. But it turns out that the ADDR_UNKNOWN value is not
used anywhere, the enum can be replaced with a bool.

Change-Id: I44f7caa649213e1aeada8d80fcf3f01794411c61