boot/{ia32/amd64}/bootstrap.ld: properly merge .bss.*
Mainly for cosmetic reasons do also consider .bss.* in the linker
script. As -ffunction-sections -ffdata-sections are passed to the
compiler, a couple of .bss.xyz input sections are created which should
be folded into a corresponding output section.

The input sections .text.* and .*data are already part of the linker
script. Add .bss.* for consistency reasons.

The resulting boot_img.x1 looks almost the same as before this change.

Change-Id: Ib862054d8e25d47d4f72cdb5c2edf34741d05391