Makerules.JDB: remove --unique when linking jdb_compound.o
The jdb_compound.o module is used similar to an .a archive with the
difference that every member of jdb_compound.o is supposed to be linked
to the final binary instead of only members which are referenced from
the binary. Hence, jdb_compound.o is generated using 'ld -r' to create a
"relocatable" module.

The --unique parameter was supposed to create separate output sections
for every (orphan) input section. (Actually all input sections are
orphan because no linker script is used for generating jdb_compound.o.)
Basically it prevents the normal merging of input sections with the same
name. This parameter seems to be superfluous because there are already
dedicated section for .text, .data, and .bss per function and data item
because -ffunction-sections -fdata-sections is used to compile objects.

Using --unique generates slightly different data but mainly generates a
lot of additional debug sections. The resulting binary looks almost
identical.

Remove this option because it seems to be superfluous (no real benefit)
and because it generates tons of annoying warnings when using with GNU
ld from the latest 2.45 branch like this:

  arm-linux-gnueabihf-ld: jdb_compound.o:
    error: unknown attributes version '

Change-Id: I1fa3c652bf5b8cd99bd64a7d05e0d7d5751b3bb5