libc/stdio: remove support for "%lc" and "ls"
The 'ps' variable is not 0 if the user specified any non-bare type, see
the 'states[]' field. The test for '(t & 15) == 3' is only relevant for
the cases in the following switch statement, so this code basically
converts
 - "%<any length modifier>c" to "%C" ('c' == 0x63), and
 - "%<any length modifier>s" to "%S" ('s' == 0x73).

As support for "%S" and "%C" was previously removed [1, 2, 3], this now
superfluous statement can be removed to improve coverage.

  [1] I9d3edcf6531ef164a01a2c17490b4170d966a8c4
  [2] I4547d909e2d4c5475ddf09516250053ab1d71b15
  [3] I9290fc19552559a32bf8b112d8595fc41a49915c

Change-Id: I4c0517328b0e4f032333f42db2de95a4395be12e