buddy_alloc: avoid external imoddiv ref from Buddy_t_base<A,B>::dump()
When compiling for older ARM platforms, the compiler might generate
references to __aeabi_idivmod() because Buddy_t_base<A,B>::dump()
evaluates (i % 10) and (i / 10) in the same expression. Change the
code to avoid the compiler from doing that.

The alternative is to implement __aeabi_idivmod() but so far this is
the only place where this function would be referenced.

Change-Id: I55e270e76c859add3dc2332b2133754616a29dad