gcc_lib: fix prototype of __ffssi2() and__ffsdi2()
The corresponding GCC documentation is confusing but according to
https://gcc.gnu.org/onlinedocs/gccint/Integer-library-routines.html,
__ffsdi2() takes an 'unsigned long' parameter while 'si' functions
take an 'unsigned int' parameter (__ffssi() not documented in above
reference).

Clang defines
  typedef int32_t si_int;
  typedef int64_t di_int;
  int __ffssi2(si_int a);
  int __ffsdi2(di_int a);

Both functions were only seen when compiling the kernel for RISC-V.
For 32-bit RISC-V, __ffssi() is used while for 64-bit RISC-V, __ffsdi()
is used.

In general it doesn't matter for these functions if the parameters is
signed or unsigned, the result is the same in both cases.

Using 'unsigned long' instead of 'unsigned long long' as parameter for
__ffsdi2() doesn't make a difference on 64-bit hosts.

Change-Id: I0589df206dc07e89c6b11a1f565abbb51d16f604