arm: set HCR.TERR and HCR.TEA
HCR.TERR and HCR.TEA shall be always set to make sure that accesses to
certain RAS feature system registers are trapped when running at EL0 or
EL1.

HCR.TERR and HCR.TEA are RES0 if FEAT_RAS is not implemented. To keep
the code simple and fast, assume that these bits are either implemented
or setting these bits is ignored by the hardware.

See ticket #CD-1591.

Change-Id: I6738361b96fec41cdf7a767220ccb844af1f9a58
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>