arm/bootstrap: Do not require start of RAM anymore
Make RAM_PHYS_BASE optional, i.e., it is not required to be given in the
BSP's Modulefile anymore. However, it can be given to link the kernel to
proper address where it is loaded such that no relocation needs to be
done during bootup.

This is a modified version without the need for new global variables,
keeping more straigt forward semantics.

Co-authored-by: Alexander Warg <alex@wargs.de>

Change-Id: Idfd807e2a8b02c4e5c6813057fa72e903abe8195
Signed-off-by: Alexander Warg <alex@wargs.de>